+++ GitHubServer1/GitHubServer.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using System.Diagnostics;
+using System.Collections.Generic;
+using System.Linq;
+using System.Net.Http;
+using System.Text;
+using System.Threading.Tasks;
+using Newtonsoft.Json.Linq;
+using Octokit;
+using System.Security.Policy;
+
+namespace GitHubServer
+{
+    public partial class GitHubServer
+    {
+        public GitHubServer()
+        {
+
+        }
+    }
+}
